Create separate configuration file for token prioritization and sorting, which can be customized easily and used throughout the app for consistency (the latter possibility not performed)
